+++ CHANGELOG.md
@@ -2,13 +2,17 @@
 
 ## 2.0.0 (2020-??-??)
 
-**Requires VSCode 1.42 or above**
+**Requires VSCode 1.44 or above**
 
 * New PlatformIO Task Explorer
   - Instant access to the Project Tasks within the VSCode Explorer
   - Grouped tasks: Generic, Advanced, PIO Remote, Platform, Custom, etc.
   - Support for PlatformIO dev-platform tasks (Program FPGA bitstream, Burn bootloader, Upload to FS, OTA Update, etc). The list of tasks depends on a particular dev-platform
   - Access to [User Custom Targets](https://docs.platformio.org/en/latest/projectconf/advanced_scripting.html#custom-targets)
+* New Project Environment Switcher
+  - Switch between project environments declared in [platformio.ini](https://docs.platformio.org/en/latest/projectconf/index.html) project configuration file
+  - Activate IntelliSense service based on the current environment
+  - Automatically generate a debugging configuration for the active environment
 * New PlatformIO Core Installer
   - Switched to the cross-platform and portable [get-platformio.py](https://github.com/platformio/platformio-core-installer) installer script
   - Use built-in portable Python 3 on Windows (it can be disabled with ``platformio-ide.useBuiltinPython`` setting)